Rename reducer again to have only date related functionalities in this one
Later most likely will have other reducers
