+++ api/front-end/src/index.tsx
@@ -8,13 +8,13 @@ import {BrowserRouter} from "react-router-dom";
 import './Magic.css'
 import {Provider} from 'react-redux'
 import {createStore, combineReducers, compose, applyMiddleware} from "redux";
-import defaultReducer from "./store/defaultReducer";
+import dateReducer from "./store/dateReducer";
 import {composeWithDevTools} from 'redux-devtools-extension';
 // import thunk from 'redux-thunk'; // Todo, install thunk if needed (lets you write async logic that interacts with the store)
 
 
 const rootReducer = combineReducers({
-  defaultReducer: defaultReducer,
+  dateReducer: dateReducer,
   /* Add other reducers */
 });
