+++ api/front-end/src/pages/Home/DateRangeSelector/DateRangeSelector.tsx
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import {ReduxPropsInterface} from "../../../store/defaultReducer";
+import {ReduxPropsInterface} from "../../../store/dateReducer";
 import {DATE_RANGE_END_DATE_SELECTED, DATE_RANGE_START_DATE_SELECTED} from "../../../store/actionTypes";
 import {connect} from "react-redux";
 
@@ -28,8 +28,8 @@ class DateRangeSelector extends Component<ReduxPropsInterface> {
 
 const mapStateToProps = (state: any): any => {
   return {
-    dateRangeStartDate: state.dateRangeStartDate,
-    dateRangeEndDate: state.dateRangeEndDate,
+    dateRangeStartDate: state.dateReducer.dateRangeStartDate,
+    dateRangeEndDate: state.dateReducer.dateRangeEndDate,
   };
 };
